tests/wpt/tests/svg/linking/scripted/a.rel-setter-01.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <title>SVGAElement.rel setter</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/linking.html#InterfaceSVGAElement"/>
  </metadata>
  <a id="test" href="a"></a>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <script><![CDATA[
    test(function() {
        var a = document.getElementById("test");

        test(function() {
            a.rel = "noreferrer"
            assert_equals(a.rel, "noreferrer");
        }, "Test anchor's rel setter");
    });
  ]]></script>
</svg>
